Multiple upgrades (#33)
Refactor scripts to common.js
Fix tooltip overflow
Fix value sorting
Update absolute sorting
Add favicon
Add save file prompt
Add default data folder
Zoom on datatable click
Fix case sensitive sorting
Add ellipses to long element name
Fix second column missing arrows
Add option to hide animations
Fix default behaviour of no dataset selected
Fix undefined default unit
Visualize selected groups on table
